Implemented the Carry Resource Entity Action
units with the carry resource entity action can pick up resources and drop them off onto a new bundle of resource or add them to an existing resource bundle. Although you can tell how much resources are in a bundle, you cannot tell what resource a unit is carrying nor how much, to be implemented at a later date.

The error codes are a bit of a mess: I wanted to force every Entity Action to implement enumerations, but I could not quite get it to work as I had wished, so there is a single enumeration that covers all error codes for all Entity Action, which I don't like very much, but it'll have to do for now.